refactor: extract PrerequisiteRow.swift from SettingsReadinessTypes.swift
Splits the `PrerequisiteRow` value struct out of
SettingsReadinessTypes.swift so that file holds only the
`SettingsReadinessStatus` enum.

Byte-identical move of lines 49-59. New file needs only Foundation
(no SwiftUI); the remaining file keeps SwiftUI for the readiness enum's
color/icon accessors.

SettingsReadinessTypes.swift: 59 → 47 lines. Tests: 667.
