+++ Sources/BugNarrator/Views/Settings/PrerequisiteRow.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+/// One row in a per-integration prerequisite checklist.
+struct PrerequisiteRow: Identifiable {
+    let title: String
+    let detail: String
+    let status: SettingsReadinessStatus
+
+    var id: String {
+        title
+    }
+}
+
