+++ Sources/BugNarrator/Views/Settings/SettingsReadinessTypes.swift
@@ -45,15 +45,3 @@ enum SettingsReadinessStatus {
         }
     }
 }
-
-/// One row in a per-integration prerequisite checklist.
-struct PrerequisiteRow: Identifiable {
-    let title: String
-    let detail: String
-    let status: SettingsReadinessStatus
-
-    var id: String {
-        title
-    }
-}
-
